apps/backend-core/requirements.txt
Django>=5.0,<5.1
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
python-dotenv
dj-database-url
pgvector
requests
pypdf
langchain-text-splitters
